refactor(qwen35): split TP executor into tp_executor/executor.rs
Move the whole TP orchestration cluster (Qwen35TpExecutor, step items TpPrefillChunkItem/
TpDecodeStepItem/TpUnifiedPlan/TpUnifiedResult, the impl, dispatch_mutating_commands,
Drop impl, plan validators, From impls, snapshot test helper) out of tp_executor.rs into
tp_executor/executor.rs. Root becomes a thin re-export: pub re-exports Qwen35TpExecutor,
pub(crate) re-exports step items so crate::tp_executor::TpPrefillChunkItem etc. still resolve
for the scheduler. tp_executor.rs 2520 -> ~724. cargo check + fmt + test --lib 95/95 + tp2
7/7 (weights + 2 GPU + NCCL) green.
